+++ src/print_test.rs
@@ -1,9 +1,27 @@
 use super::*;
-use crate::types::{ConfigSection, Task};
+use crate::types::{ConfigSection, ExecutionPlan, Step, Task};
 use indexmap::IndexMap;
 
 #[test]
-fn print_default() {
+fn get_format_type_default() {
+    let output = get_format_type("default");
+    assert_eq!(output, PrintFormat::Default);
+}
+
+#[test]
+fn get_format_type_unknown() {
+    let output = get_format_type("test123");
+    assert_eq!(output, PrintFormat::Default);
+}
+
+#[test]
+fn get_format_type_short_description() {
+    let output = get_format_type("short-description");
+    assert_eq!(output, PrintFormat::ShortDescription);
+}
+
+#[test]
+fn print_default_format() {
     let mut config = Config {
         config: ConfigSection::new(),
         env: IndexMap::new(),
@@ -14,7 +32,7 @@ fn print_default() {
     config.tasks.insert("end".to_string(), Task::new());
     config.tasks.insert("test".to_string(), Task::new());
 
-    print(&config, "test", false);
+    print(&config, "test", "default", false);
 }
 
 #[test]
@@ -29,5 +47,29 @@ fn print_task_not_found() {
     config.tasks.insert("init".to_string(), Task::new());
     config.tasks.insert("end".to_string(), Task::new());
 
-    print(&config, "test", false);
+    print(&config, "test", "default", false);
+}
+
+#[test]
+fn print_default_valid() {
+    let step = Step {
+        name: "test".to_string(),
+        config: Task::new(),
+    };
+    let steps = vec![step];
+    let execution_plan = ExecutionPlan { steps };
+
+    print_default(&execution_plan);
+}
+
+#[test]
+fn print_short_description_valid() {
+    let step = Step {
+        name: "test".to_string(),
+        config: Task::new(),
+    };
+    let steps = vec![step];
+    let execution_plan = ExecutionPlan { steps };
+
+    print_short_description(&execution_plan);
 }